restructuring tests done
