+++ crm/tests/__init__.py
@@ -1,0 +1,11 @@
+TEST_USER = {
+    'username': 'test_user',
+    'password': 'Word9876',
+}
+
+TEST_CUSTOMER = {
+    'first_name': 'Cust',
+    'last_name': 'Test',
+    'phone': 1234567890,
+    'email': 'test_cust@test.com',
+}
